test: LottoScoreBoard 객체 테스트 추가
